refactor(core): split into separate files
